feat: add a handler to the mouse
