+++ style.css
@@ -1,6 +1,7 @@
 body {
     display: flex;
     justify-content: center;
+    background-color: #ccc;
 }
 
 .wrapper {
@@ -8,8 +9,8 @@ body {
 }
 
 .textarea {
-    width: 644px;
-    height: 15vw;
+    width: 673px;
+    height: 12vw;
     border-radius: 15px;
     padding: 10px;
     outline: none;
@@ -22,6 +23,9 @@ body {
     justify-content: space-around;
     width: 667px;
     height: 14vw;
+    background-color: rgb(20, 19, 19);
+    padding: 15px;
+    border-radius: 20px;
 }
 .key {
     width: 40px;
@@ -31,6 +35,10 @@ body {
     display: flex;
     justify-content: center;
     align-items: center;
+    cursor: pointer;
+    background-color: black;
+    color: white;
+    font-style: normal;
 }
 .control {
     width: 50px;
